[roll] Roll fuchsia [userboot] VmoBuffer resize allocates page aligned sizes.
When reaching capacity this buffer would attempt to resize, up to the
minimum size to fit the string being written. Unfortunately, this was a
guaranteed fail, since in order to succeed the new size must be page
aligned.

Now we round up the size up to the nearest page boundary that contains
the appended string.

Original-Fixed: 384793954
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192155
Original-Revision: 15f16b7571e0fa5b8fece5f79477853847986936
GitOrigin-RevId: bdf00019954ad59ec962cda53a75ff0bf6342a84
Change-Id: I5e87b3be4956b241d510f2cb118946f65b461774