[roll] Roll fuchsia [kernel][crashlog] Add runtime to RAM crashlog
This change adds the runtime field to the RAM mappable crashlog's
header. This required the creation of new ram_crashlog_v1_t and
ram_crashlog_header_v1_t structures, as the addition of this field
must be done in a backward compatible way with previous versions of the
crashlog.

Note that previous versions of the kernel that use v0 crashlogs will
not be able to recover v1 crashlogs, but this is not an issue because
we do not support downgrade OTAs.

Original-Bug: 371420920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1179713
Original-Revision: 13237e858ddf883674ea426469797f689a0bbb81
GitOrigin-RevId: ac390160e160ed50fa1682dd47ffbaeb62786142
Change-Id: Ifa0e0f6d37b8457b3ff0e21b5f45be28d011426f