[roll] Roll fuchsia [vfs][cpp] Allow in-place construction of node attributes
Refactor internal wire table builder so that we can eliminate extra
copies of node attribute data on the stack. Previously, the result of
calling Vnode::GetAttributes was required to be stored on the stack, as
it was assigned to an already constructed object. It also required
unpacking the `zx::result` variant, which further complicated elision
efforts.

This should provide minor performance improvements to the GitStatus and
StatPath benchmarks for memfs and f2fs.

By instead calling Vnode::GetAttributes in the constructor of the wire
table builder, we can ensure the compiler can optimize out the extra
copy. See https://godbolt.org/z/59ExM74br for a more generic example
that demonstrates this technique. Many thanks to the reviewers of
https://fxrev.dev/1194315, the discussion there was crucial in
discovering this approach.

Original-Bug: 391921991
Test: No functional change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195918
Original-Revision: 25b1dd30a912e71e8c0094c1dd53e5af9fca10ff
GitOrigin-RevId: 9201ea9e9902d68a2c1cd755b0a3811c453308e0
Change-Id: I0b2a72f6d82857cc37bc3d61196b92832d5fe570