[roll] Roll fuchsia [bazel] add self-test to refresh-bazel-compdb
This CL adds an option to run a self-test on the commands created when
running `fx refresh-bazel-compdb`. The self-test ensures that the
compile commands are valid.

This check is not intended to be run as part of CQ due to the amount of
time it adds. The purpose of this command is to help with debugging.

Original-Bug: 392101170
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193397
Original-Revision: ad2a01b44da766c9340a334542129d79c0376be6
GitOrigin-RevId: 313091671bab1b5e662a9a35a09053536a2f57b5
Change-Id: Ife9870cc6db557c3fe4304acb005defd5251b6bb