[roll] Roll fuchsia [time][inspect] Corrects the way inspect properties are recorded
The old code used the `record_*` API which churns through inspect
memory. This is incorrect and not needed.

This change moves the creation of inspect prpoperties out of the
loop, ensuring they are rewritten, and not recreated on each
loop iteration.

While I was there, I reorganized the properties a bit as well.
Since no tooling depends on this data schema just yet, it is
low risk to change.

Original-Bug: 393633010
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196286
Original-Revision: 41e8df40eb4d36ed7eea23ff6dc4f391e34ec4d2
GitOrigin-RevId: 49923a512fbceec3e12a175d9a2d49016c43e674
Change-Id: I22191fdcb4887711699d854f6e5b9d21c8c40cfb