[roll] Roll fuchsia [testrunner] Fix profile merging when no debuginfod servers are set.
If the env var where we provide the debuginfod server URL is empty,
strings.Split() returns a list of one element which is the empty string.
So the covargs merge function was treating that as a nonempty debuginfod
server and trying to use debuginfod to merge and then failing.

Original-Bug: 394129919
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198663
Original-Revision: 5d8c2ef849c92894ee6b5db55c0012640b8eac65
GitOrigin-RevId: 99ac83df2b2827eb2bb587ed0cb3fb860499d99f
Change-Id: I50b7393d127cbdf412ed5af281ac08b5432f5631