[roll] Roll fuchsia [ffx][fho]Removing TargetInfo reference from DeviceLookup
This removes the TargetInfo reference in the DeviceLookup trait
and replaces it with a Trait defined in fho.

This trait is implemented by TargetInfoHolder in target_holders.

This is done as part of the effort to remove fidl dependencies on
the fho crate in order to improve build times of non-device subtools.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198576
Original-Revision: 6f1e5b76bce7e2d53f9193d531780a7e0838fdf5
GitOrigin-RevId: 9984ec70dfcf230c35031c8a92cb1fbd18bdde60
Change-Id: I7aad33ffde137aaa776a37371c67223449dfa551