[roll] Roll fuchsia [owners] Remove mkearney@google.com from tools/clidoc/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/mkearney@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198720
Original-Revision: 60ff2ac5efa0fa72215b2902c28c4a73c0f43011
GitOrigin-RevId: 09824b3b2213b64ded36551c6b11674c5fb97c63
Change-Id: Ic636ec5d0d44d7a6b02ccbfb791f9afef6cad6c5