[roll] Roll scudo [scudo] Refactor the secondary test (#125595)
Remove all redundant code and create a couple of structs to handle
automatic init and destruction. This replaces the test fixtures in
prepartion for passing in multiple configs for some of these tests. This
is necessary because not all of the gtest features are supported here,
and there is no easy way to create a test fixture with a template.

GitOrigin-RevId: 16a1b1030004713cf21198b2c157e55ba3cfcf39
Original-Revision: 8e022782d84ccd02f2032012f33ee47002406adb
Change-Id: Ide7fd690d22369a37aa7f571240bfa380b20dae9