[roll] Roll fuchsia [kernel] Add a runtime flag to control kernel PCI logic on arm64
This removes the WITH_KERNEL_PCIE #ifdef and adds a boot option that
controls the kernel mode PCI initialization and system calls. This
option is defined in the kernel.arm64 namespace and is only checked on
aarch64. User space PCI handling has been enabled already on x86_64 and
RISC-V and works on some arm64 platforms.

On x86_64 and RISC-V this change compiles the PCI support code into the
kernel. This code is disabled at runtime.

On aarch64, the PCI code is compiled in (as it was before) and is now
guarded by a default-true runtime flag. We can disable this flag in
configurations that do not rely on the kernel mode support and remove it
once we no longer need it. Concretely, currently the qemu-arm64
configuration only works with kernel PCI support.

Original-Bug: 42182308
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195575
Original-Revision: 0c311545f66a64b19e066c3a69c188c6e14b6ca2
GitOrigin-RevId: 4c242828d494be87164b63c189109a670d48fd1a
Change-Id: I915fce551547c7e87b99b1515cba383b743261cf