[roll] Roll fuchsia [flatland] Modify DisplayCompositor to use WireSharedClient
This is the first step in reducing the number of blocking calls in
Scenic.

While this change was mainly focused on modifying DisplayCompositor,
migrating to a WireSharedClient requires the use of wire types, which is
what the majority of this change consists of.

Original-Bug: 382704825
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198124
Original-Revision: 8e19950209808d6892dfc6734b0b7ffe76bcf1e0
GitOrigin-RevId: 157c8b86273e5c8800b1d26c19495fb82a49cc6a
Change-Id: I11766073605e05aadaf88e423eb5814cc61ba60f