[roll] Roll fuchsia [build] skipprofile attribute for excluded funcs
We currently use -fprofile-list= flag to exclude
instrumenting files that are under //prebuilt
https://clang.llvm.org/docs/UsersManual.html#cmdoption-fprofile-list

We were using the older format that is defined in
https://clang.llvm.org/docs/SanitizerSpecialCaseList.html

When older format is used, Clang adds noprofile attribute
to a function that blocks inlining if the caller and callee
have different values of this attribute as explained in
https://llvm.org/docs/LangRef.html#function-attributes.

This patch switches to a new format to skip instrumenting
a specific path by using "source:path=skip". This format
adds skipprofile attribute instead of noprofile attribute
to the functions that are under the provided path. This
attribute does not restrict inlining between functions in
the same way that the noprofile attribute does. This patch
also clarifies the comments about the usage of this new
format.

Original-Bug: 387503298
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1197420
Original-Revision: 73069b17b0b239df18eb2be55c8391b3a4941367
GitOrigin-RevId: 7da06936125695c6f208efdf7db71af6628dc3fd
Change-Id: I517714d080dac87fb52ed3ceff11b8d4097c30e5