[roll] Roll fuchsia [kernel][vm] Cleanup some lock annotations
Removes usages of AssertHeld in leaf functions with preferences to
declare a TA_REQ and, if needed, perform an AssertHeld in the caller.
Some pointer types in parameters are trivially changed to allow for
the annotations to work.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1196396
Original-Revision: 7ca781d9ae6521144c2d6bae6d8f70eb63ece80b
GitOrigin-RevId: 870546a2e65fbc75cca3909726dd3dfc01fddc45
Change-Id: I45bc8cb1d95372c3e525360beabb1cc0e134b1d8