[roll] Roll third_party/go [go] Add ability to open namespace nodes directly
To transition `ServiceConnect` to open3, we must first provide an
alternative way of opening namespace nodes with an explicit set of flags
provided. Currently, we have callers relying on the presence of rights
when using this function to open directories, however connecting to
services does not require rights.

Although connecting to a service in /svc does not require permissions,
many callers assume the presence of rights on the connection for various
operations, such as enumerating the instances of a service.

Original-Bug: 356225729
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/1200564
Original-Revision: 07b8094ea79cd6e746f252b55ef16f19272e93aa
GitOrigin-RevId: dd6e77a6d0a28a71416fe826b9bccecde2d51f31
Change-Id: I7639835521e26909b872ce0c58eb94aa506a078d