[roll] Roll fuchsia [idk][build] Enable starting sub-builds earlier
Previously, the entire main build of the core IDK collection had to
complete before sub-builds could start. This included waiting for the
conformance tests and other targets that are not built in the
sub-builds.

To avoid this, create a peer IDK collection that only runs in the
sub-builds. The sub-build rule only depends on building this collection
in the main build, so the sub-builds can be started as soon as this
subset completes.

Original-Bug: 393154095
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198647
Original-Revision: f14c12733a5c58889bb7c7afc971738c8091705a
GitOrigin-RevId: 80ac8d9a730c1f747451c32dcee296f44d707555
Change-Id: I736983926c99ee05c83fb613ce05d39e7d31240a