[roll] Roll fuchsia [packet-formats] Temporarily suppress leaks
packet-formats-test  has a leak that was uncovered by fxr/1197316.
It's likely the leak was always there but that CL improved the logging
which allowed the diagnostic to be found by tefmocheck. Util we fix it,
we can temporarily disable lsan for this test.

Original-Bug: 394935218
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201385
Original-Revision: 33254b2553ee3a12ce15a950cad4c9b509156f21
GitOrigin-RevId: adf56997eb8efc17bb1a94834209cc83ad09144b
Change-Id: Ide5acf59ce1195a18e61e6f5fd37e321ccefcdb4