[roll] Roll fuchsia [starnix] Break Arc<Kernel> cycles.
This is an important pre-requisite for dropping the Kernel
when we shut down.

However, it's quite difficult to get the teardown ordering right with
the (somewhat deprecated) `create_test_*()` APIs, so this change
opts out of running `Kernel::drop` in unit tests. As tests migrate
to the newer `spawn_kernel_*()` APIs for tests the tests will
exercise the right lifecycle paths.

Original-Bug: 295073633
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193955
Original-Revision: cbc227329457d6164ef53adfd46826080ba4424c
GitOrigin-RevId: e34730361764027808eefc815625bd4368331ca6
Change-Id: Ia6ce8596b26b8181aeef3646ade88ad9e7ba77f3