[roll] Roll fuchsia [fuchsia.io] Un-mark DeprecatedClone as `removed=HEAD`
We still have some load-bearing functionality that depends on the
existence of this method, so we should not mark it as removed until that
is fixed. Additionally, this is preventing building the IDK at HEAD.

For now we just mark the method as deprecated instead of removed.

Original-Bug: 394878428
Test: No functional change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201227
Original-Revision: 1d88a1b91bf2de60e9ff5f979f56b1bf05122b8a
GitOrigin-RevId: 33c4df7036aec18b9aac014f5b0bead9ece35945
Change-Id: I02979d8e9d447e14c6cbb15a9e5f635bf4860a57