[roll] Roll fuchsia [bt][ffx] Remove pandora from SDK partner_internal
The unstable FIDL library "fuchsia.bluetooth.pandora" was in the
"partner_internal" category because `ffx_bluetooth_tool` was in the
"partner" category due to the previous default for the template. Since
the tool is not in an IDK, remove the category from the dependencies.

This addresses one instance of an exception to "partner_internal"
stability rules.

Original-Bug: 364727245
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1194375
Original-Revision: 28621ac5b3ba5651cde447752eba82c3fa1fc05f
GitOrigin-RevId: facf86652ed6c662221e88b7bc34da995ce1be98
Change-Id: I50542663e828005a4a6e3ae9cc8b93d105df3cd3