[roll] Roll fuchsia [Starnix] Resolve socket protocol on socket creation
Socket stores `protocol` field. Previously it was always set to the
value passed to `socket()` syscall during creation. The syscall allows
`protocol` parameter to be set to 0, which case the actual protocol
used by the socket is either TCP or UDP depending on the socket type
parameter. Updated `Socket::new()` to resolve the packet type, which
ensures the correct value is stored in the struct. This will allow
to pass correct protocol value to eBPF programs executed in Starnix.

Original-Bug: 391380601
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201768
Original-Revision: 1f2292fe5d973217f3d3d678eda6d5d330998981
GitOrigin-RevId: bd17ed145b2abee817d514f016df1c3f5261d4b5
Change-Id: I1404099fd59ba11fb1177fee22a338da700db354