[roll] Roll fuchsia [flyweights] Address flakes on host.
These tests share a process on host, so we need the serial
annotation on every test case. That was missing from
utf8_and_bytestrings_share_the_cache which relies on counting items
in the global cache.

Also adds missing reset_global_cache() calls to new test cases.
Without this, host-side tests can have confusing results when one
failing test compromises the invariants of another.

Also changes the internal API for constructing a RawRepr to
require exclusive access to the cache so that reference counts
are guaranteed consistent. This is not a functional change given
the current constructors but is a useful guardrail as we refactor.

Multiply: flyweights
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203549
Original-Revision: 7834eb14cea6536c6047ba2b539ea31508877291
GitOrigin-RevId: 4475f2b3e0c93b07a02bce45867ac7f5b79c010b
Change-Id: I6ff56605e9a9d1e82d7439b1aac93363a2c10cd7