[roll] Roll llvm-project/libc [NVPTX] Make ctor/dtor lowering always enabled in NVPTX (#126544)
Summary:
Currently we conditionally enable NVPTX lowering depending on the
language (C/C++/OpenMP). Unfortunately this causes problems because this
option is only present if the backend was enabled, which causes this to
error if you try to make LLVM-IR.

This patch instead makes it the only accepted lowering. The reason we
had it as opt-in before is because it is not handled by CUDA. So, this
pach also introduces diagnostics to prevent *all* creation of
device-side global constructors and destructors. We already did this for
variables, now we do it for attributes as well.

This inverts the responsibility of blocking this from the backend to the
langauage like it should be given that support for this is language
dependent.

GitOrigin-RevId: ddc096e517eb35f7a6dfd72682af67897204ad9f
Original-Revision: 4dc3dc2ce2cb7ba01842d10fe479d9ffbfd275c7
Change-Id: I4f134c74a083df89ce59dbe8a16858dc4227e84c