[roll] Roll fuchsia [memory] Don't overcommit inspect slots
Before this change, logger.cc would allocate a fix amount of memory
for logging bucket names in the current configuration, which is
wasteful and does not print nicely (`ffx inspect show
core/memory_monitor` for instance would display a lot of empty
slots). This was done because the number of slots was not known at
logger instantiation time, but we can do better by deferring the
instantiation of the property to a time when the number of slots is
known.

Note that the number of buckets is expected to stay constant during
the entire lifetime of the logger, which is why we can get away with a
single dynamic allocation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203045
Original-Revision: 1b726b4acdaf27dfc2b629dda07f54ede215063a
GitOrigin-RevId: 7dc3162a019e95e3bcf554141677776da9de7b1f
Change-Id: I2cf78fe050b1d435d51be549af969da89bbae866