[roll] Roll fuchsia [fxfs] Enforce tree cache size limits.
This was missed when it was added, and given the relatively low usage we
didn't bother to fix it. Now the entry count has grown, and so has the
size of the entries.

Adds a test to ensure that things can age out.

Original-Bug: 393365596
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203584
Original-Revision: 298df5317c07b28ca75f665c70365eb4fe78ce5a
GitOrigin-RevId: 3c3c7393ea5fd3a5988c513f283b628b35375543
Change-Id: I837e3990aa2666faa16e1243fafa65ee23995e45