[roll] Roll fuchsia [ffx][fho] Removing re-export of ffx_writer
This removes re-exporting ffx_writer in fho. This
is done as part of refactoring the fho crate in order
to improve tool build times.

The next step is to refactor ToolIO, which is used as trait for
FfxMain, and is implemented by the Writers in ffx_writer.

Original-Bug: 392174596
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1201770
Original-Revision: bc500c71efa46f0d8f3397825f2d68fda57f7507
GitOrigin-RevId: c865e8882e6f2557813420590489141e74d73687
Change-Id: I7844c9cd52f658c390a41c9c87493d6e124f4676