[roll] Roll fuchsia [display][coordinator] Stop applying empty config on client teardown.
Client:TearDown() currently unconditionally calls ApplyConfig() at the
end of its cleanup. This applies an empty (zero-layer) display
configuration, because the cleanup empties the layer lists for the
pending configuration.

This logic is incorrect, and conflicts with the correct behavior
implemented in Client::SetOwnership().

The correct behavior when a client disconnects is to have the
Coordinator reapply the last configuration applied by the new display
owner. If there is no such configuration, or if the configuration no
longer passes the verification, the current configuration must remain
displayed.

Original-Bug: 394413629
Test: fx test //src/graphics/display
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204347
Original-Revision: 59c53608582b8ae73312bd350c30f0a5292ff4e4
GitOrigin-RevId: 72f50243f44b553e0f49fe146939d4d3d647d0c1
Change-Id: Ie53be273cd118a089e8bdfe93370972b9792a5f2