[roll] Roll fuchsia [pci] Check if device is extended (ecam) or not (cam)
This is needed in order to compute the address of device configurations.
crosvm presents a pci devices with a  "pci-host-cam-generic" compatible
string as opposed to qemu's pcie device with an "pci-host-ecam-generic"
compatible string. This uses the compatible string to decide if the
device is pci or pcie.

Original-Bug: 42182308
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204564
Original-Revision: 6eb41059857899e90d3ce13773ef2d01bc307de2
GitOrigin-RevId: 36af367e5749e6723b935dc10d6e4cdf1315f79a
Change-Id: I8720a436bd11bef9cb711abac1d21e9acf1d0bb2