[roll] Roll fuchsia [resultdb] Preserve directory structure in artifact names.
This also adds an OutputDir field to the TestDetails struct so that we
can use that to write the per-test logs from tefmocheck into and we can
omit that dir from the artifact name since it should be common among all
the artifacts for a test. It also changes the paths in the OutputFiles
to be relative to the OutputDir, so fxrev.dev/1203828 needs to go in
first to update the recipes to get the correct paths if OutputDir is
provided.

Original-Bug: 394639114
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200111
Original-Revision: d3a94e8a0d52eff79f9173916f733333fc80e76e
GitOrigin-RevId: d40013be1c5763e0ec0e4ed7bfa6f25e567e0c7a
Change-Id: I71773d8eb2c10a1d80fab3f50fc6c967015cf859