[roll] Roll fuchsia [device-watcher] Open directories with PERM_READABLE
Previously in io1, the ability to enumerate directory entries and open
child entries was unprivileged. In the current io2 rights model, we want
to constrain these to the ENUMERATE and TRAVERSE/CONNECT rights,
respectively.

To ensure we can start to enforce this, we must ensure PERM_READABLE,
which is a convenient alias including these rights, is used when opening
directories that we want to perform these operations on.

Original-Bug: 324111518
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204165
Original-Revision: 36780a91543957d7e10720744b080de5f52778f3
GitOrigin-RevId: 211e8226c81ad626230e7ae6b0e6b86813a20573
Change-Id: Ia21f0453ad74d4fe4bcdaadcb8212b96c0b4144a