[roll] Roll fuchsia [graphics][tests] Refactor conformance tests for extensibility
Refactor magma conformance tests to use a multi-component design so
that vendor specifics may be isolated in a helper component.

Removed the SDK package because it's not being used, and the SDK
can't handle testonly dependencies, which is too hard to workaround.

Conformance tests for vsi and intel will be done in a later CL.

Original-Bug: 394151761
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198863
Original-Revision: 269cc6d9478e31266bce173cd565a19e5f9186f2
GitOrigin-RevId: 5b7d6d981a3e67a97e7e7b5267adc2ebe5c7f4ad
Change-Id: I5c79ea5690ea39bf7e385a407acf61c3e5a0ef07