[roll] Roll scudo [scudo] Modify header corrupption error message (#126812)
Update the error message to be explicit that this is likely due to
memory corruption.

In addition, check if the chunk header is all zero, which could mean
corruption or an attempt to free a pointer after the memory has been
released to the kernel. This case results in a slightly different error
message to also indicate this could still be a double free.

GitOrigin-RevId: b221719d5a4b9448af6ceece51c00f1b61dcf588
Original-Revision: a8735a296f65fad4ec6527350e6667a6824f6ead
Change-Id: If0c40006a33ee43da2b9abe7eb95051e57fd4a68