[roll] Roll fuchsia [superproject] Roll scudo [scudo] Modify header corrupption error message (#126812)
Update the error message to be explicit that this is likely due to
memory corruption.

In addition, check if the chunk header is all zero, which could mean
corruption or an attempt to free a pointer after the memory has been
released to the kernel. This case results in a slightly different error
message to also indicate this could still be a double free.

GitOrigin-RevId: 7bad002603b620f425f2bc2ade483447920a7c23
Original-Revision: a8735a296f65fad4ec6527350e6667a6824f6ead
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1205599
Original-Revision: d4ccf8f6011faebb2ce4f59777a23956f20b502d
Change-Id: I6d2475a0d291c8ff76e5e3d51adcaec627ecd889