[roll] Roll fuchsia [starnix] Plumb through disk usage or approximations.
Some of the disk usage values are defaulting to zero, or doing some
wrong math on calculating the block counts.

Leaves in place the TODOs to get the right values in there when we
replace the zeroes with approximations.

Original-Bug: 367090481
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203387
Original-Revision: 474d908f2b22d925d8b056ea72c5d1777f33593b
GitOrigin-RevId: 342caa1f2888ae08c84ea43e3bded5a08358d9d9
Change-Id: I051e63e88ac9fe42b6175a084bf82a9fe1f3663f