[roll] Roll fuchsia [power] Rename tests for shutdown-shim
Shutdown-shim is unrelated to power manager anymore, and so some tests
names are adjusted.

About half of the tests are redundant now as they are equivalent to the
other half. They can be removed in the next CL as the gerrit couldn't
show a good diff.

Original-Bug: b/366035937
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1188554
Original-Revision: 4bfa121bb8ddaba22485bf8d460e26ed366c8d62
GitOrigin-RevId: 9fa3546f30841d6d4662af2da1bc588d873fa21c
Change-Id: Idaee52556222c50f26f2b9078e07261efff6d5c1