[roll] Roll fuchsia [assembly] Postprocessing script inputs are in config
The board config now lists all script dependencies
inside the config itself, as a new inputs = {} field.

This field maps the path relative to the execution
environment to the path inside the container.

This is necessary so that the board configs can be
converted to an AssemblyContainer which walks the
fields to copy files into the output directory.

Support for a postprocessing script is removed from GN,
because it is unused.

Original-Bug: 388852272
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204506
Original-Revision: 259d526c526011a8ff9dae0e7f46c0874a3ddde8
GitOrigin-RevId: 53bbbab38bc3d5a0c68a927cc67fa594343ffe0d
Change-Id: I92eed91057f8f65280eb79691f9adb7e87bb9b8c