[roll] Roll fuchsia [test][hrtimer] Improve the fidelity of the `Stop` method
The hrtimer API promises that a StartAndWait will return Canceled
for a stopped timer. The fake hrtimer used to not do that.

The new code tries to ensure that the cancelation process happens
in a way that is specified by the hrtimer API. The process is
less efficient than what hardware does, but should be compatible.

Multiply: starnix-tests
Original-Bug: 373731551
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206664
Original-Revision: fd0da9cbb4f53ea98d67d80e4695bd5dd6741ab2
GitOrigin-RevId: a8e2f379e9d96e204380bea73b0cdda1251317f4
Change-Id: Ic1aa9327f2c505f57e964888b0afb98318882f28