[roll] Roll fuchsia [kernel][dpc] Make Dpc's func_ and arg_ members const
Make Dpc's func_ and arg_ data members const in order to make it
easier to reason about the class's thread-safety.

Note, for this to happen, we had to remove a check from
test_dpc_queue.  That check was actually testing something that wasn't
guaranteed to be true (that it's safe to destroy/overwrite a Dpc once
it has begun executing, rather than before it has finished executing).

This is a refactoring change with no intended change in behavior.

Original-Bug: 395669867
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203446
Original-Revision: 5e98ebaa08df807be41cfa7746e84b0c4b1d7f40
GitOrigin-RevId: eb31ec211b2f212b287de48118c90e6887f99af7
Change-Id: I0211faebdef064ec40be7193823a2037a94fe89c