[roll] Roll fuchsia [owners] Remove zhiyichen@google.com from src/connectivity/wlan/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/zhiyichen@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207703
Original-Revision: dae1f008243fb25450e6afef9418bcfa491698d0
GitOrigin-RevId: 1d9adfdbb7d3d66ef14b5d5e30529ac5d3cb3850
Change-Id: Icc4129f86a79557c0bd811f70c7440b719ba326b