[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Produce standard-compliant header guard macros in hdrgen (#127356)
Macros starting with alphabetic characters such as "LLVM" are in
the application name space and cannot be defined or used by a
conforming implementation's headers.  This fixes the headers that
are entirely generated, and the __llvm-libc-common.h header to
use a conforming macro name for the header guard.  That is, it
starts with "_LLVM_LIBC_" instead of "LLVM_LIBC_", as identifiers
starting with an underscore followed by a capital letter are in
the name space reserved for the implementation.

The remaining headers either will be fixed implicitly by removal
of their custom template files, or will need to be fixed by hand.

GitOrigin-RevId: b731ab270e5c053c68e313d68acc0ec557744840
Original-Revision: c137c112faf4b47cc446b54a16d7c532819b874f
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1208044
Original-Revision: b4fdddcbb23077e99e4d96eea22792cc08de2b95
Change-Id: Ide4a9e21893082d3fa5fd043cfb88230ce94410f