[roll] Roll fuchsia [unwinder][CFI] Try both debug_info and binary files
The CFI unwinder has the privilege of potentially having two sources of
information. We prefer using the "debug_info" ELF file when it's
available since this will contain additional debugging information like
the debug_frame section.

When that fails or is not present, fall back to the binary object which
may be more readily available and/or have eh_frame sections which are
also effective for CFI unwinding.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195600
Original-Revision: 28174d106faa45510c4eb49767a7e59022fdc2a3
GitOrigin-RevId: 903d0fee836e11cf312b4e8b5acacffa247b7063
Change-Id: I859f5898395c54f722e3eda8b6f5a32f5cba0560