[roll] Roll fuchsia [kernel][vm] Cancel request in test
Corrects the unittest to cancel any outstanding requests between
iterations. Although the request does not need to be waited on for the
test, failing to wait will trigger an assertion if a second request
needs to be generated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1208167
Original-Revision: daddc22ec7904505e59c7e4f8a89b914f68d7c00
GitOrigin-RevId: 4caa72098f72665973bfc9bdf9aa6760d2726efd
Change-Id: I4d487904e5c98b84ff4c0ab933309e118ac2020a