[roll] Roll fuchsia [netfilter] Box errors to reduce stack usage
We are trying to roll out the 'large_result_err' lint, which reports
result errors that are larger than 128 bytes, which can impact
performance.

This modifies the errors to be boxed, which hopefully improves the
performance in non-error paths.

Original-Bug: 324167674
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207246
Original-Revision: cfb4c5ff49560733f42fee8b182c5220ae3daba1
GitOrigin-RevId: f66b628d6c6f17f4f67af007ea930a6231c0eca3
Change-Id: Ic0882fb60cb36a1dfae327fc9651920d642af232