[roll] Roll fuchsia [modp_b64] Add OWNERS file.
Add myself as an owner so we can add an extra target to this library in
the following patch. The extra target consist on a forced c-copmpiled
version of this library, to made it usable in non std compatible
contextgs (e.g. physboot).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207344
Original-Revision: 75a3b313c7fe3e75060e1fe614e9cf39ae3206dc
GitOrigin-RevId: 735c8ffd90a42055899e618b9e0a2296f0731880
Change-Id: I3bb3e7404376e95744ca15742907a7ec56cd84f9