[roll] Roll fuchsia [gidl][fuchsia-controller] Add OWNERS for fuchsia-controller backend
Both awdavies@ and chcl@ are building fidlgen_python and will be making
changes to the fuchsia-controller GIDL backend as we discovers parts of
the backend that should change. For example, these is a follow-up change
to avoid using the Python keyword `bytes` as a variable in tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202424
Original-Revision: 13bb27312970aef7d12b7f4dd3b88dbfae8dcf6c
GitOrigin-RevId: 334db65810abfbefab5a957b1e9d26455002eecf
Change-Id: I332462eee127aecf04103c67a3f26deea4d29496