[roll] Roll fuchsia [bazel][ffx] Add SDK manifest to ffx product create
This adds the explicit configuration to use the SDK manifest when
running `ffx-product create`. This is needed since the SDK version is
read from the SDK when building the product bundle.

This change is in preparation for not implicitly reading the
SDK manifest in-tree, which will be done in a follow-on CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1209104
Original-Revision: 5db36ba5b5355427e02157d77910090f4431df9c
GitOrigin-RevId: 4e358ecedf343fb349600bf88fbd180124bb8a60
Change-Id: I66ec18bfeefe1062acde3eb5a5daf08e9d516914