[roll] Roll fuchsia [netstack] Encode skips in ANVL test expectations
Encode test cases that should be skipped or that will be skipped by
ANVL in expectations as Skip and AnvlSkip respectively.

Delete the test expectations for DHCPv6 client PD against NS2 because
the suite is only run against NS3 due to license contention in Infra,
and the presence of expectations now implies that a suite is run.

A follow-up change in the network-conformance repo will omit the
skipped test cases when running the full suites instead of encoding
the skips in GN.

Original-Bug: 355222753
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204745
Original-Revision: 633151feb6447ec4105ef888887fd662189f6050
GitOrigin-RevId: d6c64e6add060aec9e8fc015c0c631545584e6b8
Change-Id: I9ba6b555cb5178934114074880b590ce2c0381fa