[roll] Roll fuchsia [kernel][vm] Thread blocked in usercopy can be killed
Fixes a bug where a thread blocked on a page request from a usercopy
could end up busy looping on receiving a kill signal. Since we do not
return out to user mode, the kill signal cannot be processed in the same
way as a user mode page fault, where we pretend that the page fault
succeeded.

This was a problem only for usercopy with inline fault resolution. Added
a test that does not terminate without this fix. Also added a test for
the usercopy variant that captures faults (this worked fine previously).

Test: core-pager-process
Run-All-Tests: true
Original-Bug: 42084841
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1205447
Original-Revision: 07d91adc3d435a37f306dc789f34c8a9020c96f4
GitOrigin-RevId: d5f217331adf62a28de9aec00145dff28e29f4f2
Change-Id: I941282fd7ac4ab7328edaebdcba1ae620a4ec55c