[roll] Roll fuchsia [owners] Remove anmittal@google.com from sdk/testing/sl4f/client/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/anmittal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180364
Original-Revision: 89cd9f6fa8d9d89273006bd9c26a3c9c9382d174
GitOrigin-RevId: 42d858a0c85a3ce7157e8c21c56bfa4656133655
Change-Id: I8cacbf34199d43b752cc59f90158c1520ef5b5fb