[roll] Roll fuchsia [owners] Remove anmittal@google.com from sdk/fidl/fuchsia.testing.runner/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/anmittal@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1180361
Original-Revision: 349aab0bef337fb07f8c4c3bb01be6df4764449a
GitOrigin-RevId: d688c5da38358cd1a73ff9f74a1581b842ce290c
Change-Id: I184041bb8b55b542abc7bce4cbd0ae3a7b7dfa35