[roll] Roll fuchsia [kernel][ktrace] Improve message flow correlation.
Improve message flows to automatically correlate channel calls end-
to-end and simplify post processing to correlate non-call messages with
transaction ids.

- Change flow id generation to use the transaction id, when specified,
  as part of the hash. This ensures that reads and writes to the same
  channel with the same transaction id generate the same flow id.

- Emit flow steps for the channel read and write operations in a channel
  call flow to connect them end-to-end.

To aid in offline correlation, bit 31 of the flow id is set when the
associated message has a non-zero transaction id and bits [0 .. 30]
contain a hash of the transaction id and channel koid, such that the
hash value is consistent for a (txid, koid) pair.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207850
Original-Revision: 8466b97781c85618b34a3969f05ed59a24ee9ba8
GitOrigin-RevId: 8f9b9281847c48a8263a21acca00f0cd15d67219
Change-Id: I764e23330eb6851d4eb166858a5eeb90b03af13c