[roll] Roll fuchsia [nanohub][starnix] Add nanohub_socket_file for sysfs EOF
This is intended to be a stop-gap solution to allow us to use sockets
for sysfs endpoints with predictable behavior. The consumers of our
interface expect to receive an EOF (or syscall or read returning 0)
after their first read containing the full payload. We are able to send
the responses in a single datagram, so this solution would carry us
while a more robust interface is developed.

Adds additional logic to underlying test to ensure EOF behaves as
expected.

Test: starnix_nanohub_module_test
Original-Bug: 391691908
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200726
Original-Revision: c16f12590e9bcd5aa270c0172da86849b6885b18
GitOrigin-RevId: 08275a3aae797c5158b9ec1bc774f544c5b68307
Change-Id: Ie8f779dd5c0facc07c2c9acf92811062c037538d