[roll] Roll fuchsia [ffx] Add trait for ffx test execution.
This adds a FfxExecutor trait that allows for commands that are not just
Isolate to run a list of commands in sequence and evaluate their success
cases.

Currently the libraries are not used outside of the emulator integration
test code, which is not in any known build paths.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210308
Original-Revision: a9773095a7761493f9b053fa6e2b4bdf38ce3e2e
GitOrigin-RevId: 91e24b698082086dc1b4544ed3dddd40a7437ca0
Change-Id: I0ab0e80080dbd285552452ee75419f6b23ca1cbb