[roll] Roll fuchsia [storage] BlobWriter prints checksum on corruption error
To help determine where the corruption is occurring, when BytesReady
fails with IO_DATA_INTEGRITY BlobWriter now prints the total number of
bytes written to the VMO so far and the crc32 checksum of those bytes,
which will appear in the pkg-resolver's syslog.

This should not affect production devices because BlobWriter is not
used in those configurations.

Bench tested that the printed checksum matches that of the crc32 CLI
tool.

Original-Bug: 378158208
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211134
Original-Revision: c5a668e559d6a6aac2c4950f7cc202f4b1a59b29
GitOrigin-RevId: 3cf28951583d5af26d9718161dbfbdd743c9290f
Change-Id: I4bd8711eb1da6deb7a34b8abc401e231cbc65391