[roll] Roll fuchsia [fidl][ffx] Apply category markers to all FIDL sub-targets
Specifically, apply it to the FIDL library on the FIDL toolchain, which
all bindings types will inherit from `fidl_ir_target`. This catches uses
of FIDL libraries within dependencies of ffx rather than just use of
them via the Rust bindings. It also makes it possible to use the markers
in contexts other than ffx.

Update the allowlist to include additional dependencies caught by the
now more thorough checks.

Note: This could catch uses of FIDL libraries or types internal to a
component such that they do not pose a compatibility concern. However,
this should be rare and can be handled if/when it is a problem.

Original-Bug: 314822328
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207239
Original-Revision: 5de416831d06d66f062f82bb6ccc4db3f632cf28
GitOrigin-RevId: 27cfb541b75f484a61f19fafdec88457b5f8d7af
Change-Id: I39671e529b2dfe575a8bfc736dc74b77acfb521a