[roll] Roll fuchsia [lacewing][ffx]: Support --net user emulators
This change allows the use of --net user emulators by setting the local
mobly driver to always use the device ip instead of using the target
name. This will mean that users will have to be aware of rebooting the
device or any change to the device ip. We have a warning in place and
will revisit if this becomes an issue.

Original-Fixed: b/394940615
Original-Fixed: b/366067471

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206705
Original-Revision: 3d4a7ea11d8b3560259bab1aebf21938f3befded
GitOrigin-RevId: d9f850624eecc13021e4bbf8d79af553e7a43edc
Change-Id: Iba64e8f2250373938a3a271de2ec2f8ccdb750ee