[roll] Roll fuchsia [fidl][build] Clarify check for internal zx library
When a target name does not start with a colon, GN's
`get_label_info()` adds an extra directory to the result. In this case,
that real directory exists and is a different library. Specify a colon
so that the literal string reflects the actual library location.

Original-Bug: 325669391
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211888
Original-Revision: cd5ff7f9392dee13837fb37caf01cbde2f3eba36
GitOrigin-RevId: c03f244e91157a68b70665f4379d2b119e957642
Change-Id: Ibfe6015fda1d037d94c0179c63d703090135dec9