[roll] Roll fuchsia [ffx][sdk] Allow specifying SDK manifest path
This change refactors SdkRoot to allow specifying the relative
path to the SDK manifest within the SDK root. Previously, it was
hardcoded to be 1 of 2 possible paths.

The previous behavior is maintained if the manifest path is None,
so we can adopt being explicit about the in-tree manifest gradually.

Original-Bug: 397989792
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211164
Original-Revision: f37ae33be55bb5b16d36d15f15a86164427885ad
GitOrigin-RevId: d3b702a477791f0db7963565093e590b5fa805d5
Change-Id: I7c424f275710c48abb2a18a3a04d4195cc372445