[roll] Roll fuchsia [kernel][vm] Split RangeChangeUpdateLocked
Splits the existing range change update into two separate methods, one
that just operates on the current node, and one that just operates on
the child nodes. When operating on the self node the caller is
additionally required to indicate whether they intend to invalidate the
children later, or if they assume that they have no children.

Performing this split is a no-op in this commit and later commits in
the fine grained locking transition will take advantage of it.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1209206
Original-Revision: 26d53de526dae973a067826cb68c90f593948a6d
GitOrigin-RevId: 5dd96da1300efaa92604d4293d507244a8ae48b6
Change-Id: Iaa8285d3bce794f0cf392d08c1e034b2ce640a5a