[roll] Roll fuchsia [unwinder] Clarify CFI loading code
This change clarifies the usages of both which style of CFI we're
reading. The eh_frame section contains "exception frames" which are
comprised of CFI records very similar, but not the same as, CFI records
contained in the debug_frame section. Within each of these is the
possibility that the CFI records are in 64 or 32 bit format, depending
on various compiler options.

It should now be more obvious to readers at a glance what kind of CFI
format we're reading from a given module. There should be no observable
changes to behavior.

Original-Bug: https://fxbug.dev/292107156
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195603
Original-Revision: c2a6b9ec2941e6af53e5e031b10742e5cc10f6d7
GitOrigin-RevId: 00d2abf10638aee75b89480d6fd0de45d5229c18
Change-Id: I8940c89f578241143f837aa4219a2d6d3175e968