[roll] Roll fuchsia [kernel][vm] Simplify range change reasoning
Moves the range change update in ZeroPagesLocked out of the if into
the main body of the method. Guarding the range change by the if is
correct, since if can_decommit_zero_pages is false we will not be able
to remove any pages or otherwise modify the page list in a way that
would require a range change update, but this correctness is subtle.

Given that performing a zero range on a contiguous vmo is a very
uncommon operation, performing an additional unmap in this case is
worth the additional clarity.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213765
Original-Revision: ea22eb6d45454e45e8106044e4a91d5fed73bd41
GitOrigin-RevId: 939739f7efad1a8769403545304d1da1c65a272a
Change-Id: Iae16c9607877d9fdf52291cf294638bc2399dcde