[roll] Roll fuchsia [flash-kola] Only limit to one fastboot device if ANDROID_SERIAL unset.
This check is a useful guardrail for cases where it's not clear the
user knows which device they are flashing.

This environment variable tells fastboot which target to use.
In the absence of proper argument parsing this is the best indication
we have that a caller of this script knows which fastboot device
they'd like to flash.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214090
Original-Revision: 31195bbfe8306acba8363987ceb4f77e8ab0ef6a
GitOrigin-RevId: 3eb60ca710ff3fa5707a806394e3f8de2ec3d49f
Change-Id: I939dda273c88168f1d258800d20db9c759466668