[roll] Roll fuchsia [display][coordinator] Make IdMappable ID immutable and private.
We hope to remove IdMappable in the long term. However, for the
short-to-medium term, we'll probably have to live with it.

This CL addresses the biggest issues in IdMappable -- the data member
that serves as the map key is both public and mutable.

1. Public data members go against the C++ style guide.

2. Most map implementations (including the ones in fbl) assume that an
   object's key remains immutable while the object is inserted in the
   map. Making the map key immutable is a straightforward way to ensure
   that the assumption above is true.

Having an immutable data member makes the IdMappable class non-copyable
and non-movable. The C++ style guide requires stating these aspects
explicitly.

Original-Bug: 336394440
Test: fx test //src/graphics/display
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211122
Original-Revision: 37f444decd03c1c931efa719f41971d7a72d53c8
GitOrigin-RevId: 7a92ba6e7e09e24bd40158a261eaae78b290bfed
Change-Id: Ic2a19e240e8cff97a8edd533232274670420193d