[roll] Roll fuchsia [kernel] End phys handoff earlier
This change moves the end of physboot handoff from userboot init to
just before kernel shell init. The reason there is that we want kernel
state to be 'finalized' before we run any kernel scripts. In particular,
we want this to be the case before we run in-kernel unit-tests.

For example, hot-plug functionality was discovered to break after
hand-off end as we would previously free identity map page tables at
this stage; however we were actually still relying on for later hot-plug
secondary CPU bootstrap. Because we ran the unit tests before ending the
hand-off phase, this breakage was silent. This change prevents that from
happening again.

Original-Fixed: 398895878
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1212969
Original-Revision: 1366f8ea3c68d0c60843836a70d16ca5a5993b94
GitOrigin-RevId: 2b2c8acc65a84afe99d8d62b9fc82462b6211323
Change-Id: I09eee99804e4fd460cfa07fdbd6b6b244ae86ba4