[roll] Roll fuchsia [ffx][sdk] Do not automatically load sdk manifest in-tree
This changes the initialization of the sdk used by ffx so in-tree,
it does not automatically try to load //$root_out_dir/"sdk/manifest/core"
as a sdk manifest. Most use cases of ffx in-tree work without needing
the sdk loaded.  In those situations where the sdk is needed, for
example tests that non-ffx tools, an sdk module is needed to be
specified in the configuration or on the command line.

The effect of this change is to allow ffx commands that do not
need the SDK to run without having to provide the SDK. This is done
by deferring the error until the sdk is accessed vs. validating the
sdk manifest when ffx starts.

NO_IFTTT="Changing the value, not the structure of the data"

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207610
Original-Revision: ce94345ef575279cd27f5833783a132b0cea39eb
GitOrigin-RevId: f2f2878d41290d6b02bec5412123917de661424c
Change-Id: Ib0edb0f3e9f6a6844c4f7d02cfac17fcb2406b1a