[roll] Roll fuchsia [setui][healthchecks] Add health check server to setui
Updates the cml to expose the health check to the framework
and adds the fidl service to the apropos service directory..

If the setui environment is able to be created, then we know
that all of the agents have responded without error. If the
environment fails, then setui reboots. An inability to
get a response from the component for its health check status
is treated as a failure. So, if the health check protocol is
able to be queried, we have enough information to deem the
component healthy :)

Tested: fx test //src/settings

Original-Bug: 397979575

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210344
Original-Revision: 99d00dda175a24c924106da17a817f4f6c1196fc
GitOrigin-RevId: acd457a50eea1ea5efe23f10705e0447627cfbb9
Change-Id: Ieab15399169e8f8c4fe25c30658ca3edc49f4e11