[roll] Roll fuchsia [display][coordinator] Move hotplug processing to client dispatcher.
This is a step towards simplifying the threading model for data
structures such as DisplayInfo, DisplayConfig, Image, and Layer.

This change may seem like a step backwards in responsitivity, as it will
cause the client dispatcher to become unresponsive while a display's
EDID is processed, which may entail multiple configuration checks. This
is not an issue for the following reasons.

1. In the medium term, all supported display hardware lacks EDID, so
   this change will not cause a regression.

2. In the long term, we're planning to move EDID processing to a
   dedicated panel driver.

Original-Bug: 336394440
Test: fx test //src/graphics/display
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210125
Original-Revision: 047c556956824df2c1601c0a390f5183dbd166cc
GitOrigin-RevId: da3ccb3cf4fc77f69b2fe7637deb31efa555fdb2
Change-Id: I0450604922ebcb349ba38e5cb35f3affad834bc0