[roll] Roll fuchsia [kernel][vm] VmEnumerator is given Guard
Extends the EnumerateChildren method to pause enumeration during
callbacks, allowing it to pass the Guard in. This allows the callbacks
to temporarily drop the lock if desired.

While changing the interface to support the Guard parameter, the return
type is also changed to a zx_status_t to be more consistent with other
callback iteration patterns.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1208171
Original-Revision: c583be573a81ad7c3f2c42ae92a12c43b41e5371
GitOrigin-RevId: 7efb6e76e96bf3d9d0f2d8badb280ce25b64b2ac
Change-Id: Idcfd8c3c6cb55653f651076e920f5096164ae1ae